Move code from __init__ to compute
